+++ tests/Header/ContentLocationTest.php
@@ -6,7 +6,6 @@ namespace Tests\Innmind\Http\Header;
 use Innmind\Http\{
     Header\ContentLocation,
     Header,
-    Header\LocationValue
 };
 use Innmind\Immutable\Sequence;
 use Innmind\Url\Url;
@@ -16,18 +15,12 @@ class ContentLocationTest extends TestCase
 {
     public function testInterface()
     {
-        $h = new ContentLocation(
-            $av = new LocationValue(Url::of('/foo/bar')),
-        );
+        $h = ContentLocation::of(Url::of('/foo/bar'));
 
         $this->assertInstanceOf(Header::class, $h);
         $this->assertSame('Content-Location', $h->name());
         $v = $h->values();
         $this->assertInstanceOf(Sequence::class, $v);
-        $this->assertSame($av, $v->find(static fn() => true)->match(
-            static fn($first) => $first,
-            static fn() => null,
-        ));
         $this->assertSame('Content-Location: /foo/bar', $h->toString());
         $this->assertSame('/foo/bar', $h->url()->toString());
     }